fs: explicit casts in GBIT macros to avoid overflows in signed types
complaints from asan

+++ fs/fs.h
@@ -67,11 +67,11 @@ void	sysclose(File*);
 #define	GBIT8(p)	(((uchar*)(p))[0])
 #define	GBIT16(p)	(((uchar*)(p))[0]|(((uchar*)(p))[1]<<8))
 #define	GBIT32(p)	(((uchar*)(p))[0]|(((uchar*)(p))[1]<<8)|\
-				(((uchar*)(p))[2]<<16)|(((uchar*)(p))[3]<<24))
+				(((uchar*)(p))[2]<<16)|((uint)(((uchar*)(p))[3])<<24))
 #define	GBIT64(p)	((u32int)(((uchar*)(p))[0]|(((uchar*)(p))[1]<<8)|\
-	(((uchar*)(p))[2]<<16)|(((uchar*)(p))[3]<<24)) |\
+	(((uchar*)(p))[2]<<16)|((uint)(((uchar*)(p))[3])<<24)) |\
 	((uvlong)(((uchar*)(p))[4]|(((uchar*)(p))[5]<<8)|\
-	(((uchar*)(p))[6]<<16)|(((uchar*)(p))[7]<<24)) << 32))
+	(((uchar*)(p))[6]<<16)|((uint)(((uchar*)(p))[7])<<24)) << 32))
 #define	PBIT8(p,v)	do{(p)[0]=(v);}while(0)
 #define	PBIT16(p,v)	do{(p)[0]=(v);(p)[1]=(v)>>8;}while(0)
 #define	PBIT32(p,v)	do{(p)[0]=(v);(p)[1]=(v)>>8;(p)[2]=(v)>>16;(p)[3]=(v)>>24;}while(0)
